requirements.txt
anthropic>=0.25.0
langchain-anthropic>=0.1.0
langchain>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
qdrant-client>=1.7.0
httpx>=0.27.0
pyyaml>=6.0.0
streamlit>=1.30.0
pypdf>=4.0.0
openai>=1.0.0
lancedb>=0.1.0
fastapi>=0.110.0
uvicorn>=0.27.0
